final adjust

+++ examples/readRaw/readRaw.ino
@@ -4,7 +4,7 @@ const byte doutPin = 2;
 const byte sckPin = 3;
 
 phaHX711 hx711(doutPin, sckPin);
-double result;
+long result;
 
 void setup() {
   Serial.begin(9600);
